% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_graph.R, R/api.R
\name{api_fields}
\alias{api_fields}
\title{Valid fields available for use in API calls}
\usage{
api_fields()

api_fields()
}
\description{
Valid fields available for use in API calls

Valid fields available for use in API calls
}
