% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_graph.R, R/api.R
\name{get_author_details}
\alias{get_author_details}
\title{Retrieve author information}
\usage{
get_author_details(author_id)

get_author_details(author_id)
}
\arguments{
\item{author_id}{string with author identifier}
}
\value{
list representing author object

list representing author object
}
\description{
This function retrieves Semantic Scholar data for
an author given the S2Author identifier

This function retrieves Semantic Scholar data for
an author given the S2Author identifier
}
\details{
Example of Accessible Paper Identifiers:
\itemize{
\item S2 Author ID : 1741101
}

Limitations:
Can only return up to 10 MB of data at a time.

Example of Accessible Paper Identifiers:
\itemize{
\item S2 Author ID : 1741101
}

Limitations:
Can only return up to 10 MB of data at a time.
}
\examples{
\dontrun{
 get_author_details(1741101)
 }
\dontrun{
 get_author_details(1741101)
 }
}
