% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_graph.R, R/api.R
\name{get_paper_details}
\alias{get_paper_details}
\title{Retrieve paper information}
\usage{
get_paper_details(identifier, include_unknown_refs = FALSE)

get_paper_details(identifier, include_unknown_refs = FALSE)
}
\arguments{
\item{identifier}{string with identifier}

\item{include_unknown_refs}{logical, Default: FALSE}
}
\value{
list representing S2 paper object

list representing S2 paper object
}
\description{
This function retrieves Semantic Scholar data for a paper
given its identifier

This function retrieves Semantic Scholar data for a paper
given its identifier
}
\details{
Example of Accessible Paper Identifiers:
\if{html}{\out{<sha>}} - a Semantic Scholar ID, e.g. 649def34f8be52c8b66281af98ae884c09aef38b
CorpusId:\if{html}{\out{<id>}} - a Semantic Scholar numerical ID, e.g. 215416146
DOI:\if{html}{\out{<doi>}} - a Digital Object Identifier, e.g. DOI:10.18653/v1/N18-3011
ARXIV:\if{html}{\out{<id>}} - arXiv.rg, e.g. ARXIV:2106.15928
MAG:\if{html}{\out{<id>}} - Microsoft Academic Graph, e.g. MAG:112218234
ACL:\if{html}{\out{<id>}} - Association for Computational Linguistics, e.g. ACL:W12-3903
PMID:\if{html}{\out{<id>}} - PubMed/Medline, e.g. PMID:19872477
PMCID:\if{html}{\out{<id>}} - PubMed Central, e.g. PMCID:2323736
URL:\if{html}{\out{<url>}} - URL from one of the sites listed below, e.g. URL:https://arxiv.org/abs/2106.15928v1

URLs are recognized from the following sites:
semanticscholar.org
arxiv.org
aclweb.org
acm.org
biorxiv.org

https://api.semanticscholar.org/api-docs/graph#tag/Paper-Data/operation/get_graph_get_paper

Example of Accessible Paper Identifiers:
\if{html}{\out{<sha>}} - a Semantic Scholar ID, e.g. 649def34f8be52c8b66281af98ae884c09aef38b
CorpusId:\if{html}{\out{<id>}} - a Semantic Scholar numerical ID, e.g. 215416146
DOI:\if{html}{\out{<doi>}} - a Digital Object Identifier, e.g. DOI:10.18653/v1/N18-3011
ARXIV:\if{html}{\out{<id>}} - arXiv.rg, e.g. ARXIV:2106.15928
MAG:\if{html}{\out{<id>}} - Microsoft Academic Graph, e.g. MAG:112218234
ACL:\if{html}{\out{<id>}} - Association for Computational Linguistics, e.g. ACL:W12-3903
PMID:\if{html}{\out{<id>}} - PubMed/Medline, e.g. PMID:19872477
PMCID:\if{html}{\out{<id>}} - PubMed Central, e.g. PMCID:2323736
URL:\if{html}{\out{<url>}} - URL from one of the sites listed below, e.g. URL:https://arxiv.org/abs/2106.15928v1

URLs are recognized from the following sites:
semanticscholar.org
arxiv.org
aclweb.org
acm.org
biorxiv.org

https://api.semanticscholar.org/api-docs/graph#tag/Paper-Data/operation/get_graph_get_paper
}
\examples{
\dontrun{
 get_paper_details("fb5d1bb23724d9a5a5eae036a2e3cf291cac2c1b")
 }
\dontrun{
 get_paper_details("fb5d1bb23724d9a5a5eae036a2e3cf291cac2c1b")
 }
}
