% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{s2_api}
\alias{s2_api}
\title{Endpoint used for requests to Semantic Scholar API}
\usage{
s2_api()
}
\description{
When environment variable SEMANTICSCHOLAR_API is set to a valid API key this
value differs as endpoints allowing higher rate limits are being used.
}
