% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{S2_attribution}
\alias{S2_attribution}
\title{Attribution}
\usage{
S2_attribution()
}
\description{
Use this attribution whenever data from the API is publicly displayed
}
\details{
Semantic Scholar provides a RESTful API for convenient linking
to Semantic Scholar pages and pulling information about individual records
on demand. When publicly displaying data from this API,
please incorporate the Semantic Scholar name and logo and point back to
Semantic Scholar at https://www.semanticscholar.org/ with
a utm_source=api UTM parameter.
}
