% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{S2_author}
\alias{S2_author}
\title{Retrieve author information
This function retrieves Semantic Scholar data for
an author given the S2Author identifier}
\usage{
S2_author(S2AuthorId)
}
\arguments{
\item{S2AuthorId}{string with author identifier}
}
\value{
list representing author object
}
\description{
Retrieve author information
This function retrieves Semantic Scholar data for
an author given the S2Author identifier
}
\details{
Example of Accessible Paper Identifiers:
\itemize{
\item S2 Author ID : 1741101
}
}
