% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{S2_author2}
\alias{S2_author2}
\title{This function retrieves Semantic Scholar data for
an author given the S2Author identifier}
\usage{
S2_author2(
  S2AuthorId,
  details = c("papers"),
  offset = 0,
  limit = 10,
  fields = NULL
)
}
\arguments{
\item{S2AuthorId}{string with author identifier}

\item{details}{one of "authors", "citations" or "references"}

\item{offset}{integer paging offset}

\item{limit}{integer paging length}

\item{fields}{extra fields to include, for example "title,authors"}
}
\value{
list representing author object
}
\description{
This function retrieves Semantic Scholar data for
an author given the S2Author identifier
}
\details{
Example of Accessible Paper Identifiers:
\itemize{
\item S2 Author ID : 1681232
}
}
\examples{
\dontrun{
 S2_author(1681232)
 S2_author2(1681232, fields="affiliations,aliases,externalIds")
 S2_author2(1681232, fields="paperId,externalIds", details = "papers", offset=0, limit = 100)
 }
}
