% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{S2_paper}
\alias{S2_paper}
\title{Retrieve paper information
This function retrieves Semantic Scholar data for a paper
given its identifier}
\usage{
S2_paper(identifier, include_unknown_refs = FALSE)
}
\arguments{
\item{identifier}{string with identifier}

\item{include_unknown_refs}{logical, Default: FALSE}
}
\value{
list representing S2 paper object
}
\description{
Retrieve paper information
This function retrieves Semantic Scholar data for a paper
given its identifier
}
\details{
Example of Accessible Paper Identifiers:
\itemize{
\item S2 Paper ID : 0796f6cd7f0403a854d67d525e9b32af3b277331
\item DOI : 10.1038/nrn3241
\item ArXiv ID : arXiv:1705.10311
\item MAG ID : MAG:112218234
\item ACL ID : ACL:W12-3903
\item PubMed ID : PMID:19872477
\item Corpus ID : CorpusID:37220927
}
}
