% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{S2_paper2}
\alias{S2_paper2}
\title{This function retrieves details about a paper's authors}
\usage{
S2_paper2(
  identifier,
  details = c("authors", "citations", "references"),
  offset = 0,
  limit = 10,
  fields = NULL
)
}
\arguments{
\item{identifier}{string with keywords to search for}

\item{details}{one of "authors", "citations" or "references"}

\item{offset}{integer paging offset}

\item{limit}{integer paging length}

\item{fields}{extra fields to include, for example "title,authors"}
}
\value{
list representing paper objects
}
\description{
This function retrieves details about a paper's authors
}
\examples{
\dontrun{
 S2_paper2(identifier = "649def34f8be52c8b66281af98ae884c09aef38b")
 }
}
