% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{S2_ratelimit}
\alias{S2_ratelimit}
\title{Rate limit for API calls}
\usage{
S2_ratelimit()
}
\description{
The minimum interval in seconds between requests to the API.
}
\details{
When environment variable SEMANTICSCHOLAR_API is set to a valid API key this
value can be approx 0.01 (for 100 requests per second) but if no API key is
used, the default value will be around 3.5 (allowing a maximum of 100 requests per 5 minutes)
}
