% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_graph.R, R/api.R
\name{search_papers}
\alias{search_papers}
\title{This function retrieves Semantic Scholar data for
a keyword search}
\usage{
search_papers(keyword, offset = 0, limit = 10, fields = NULL)

search_papers(keyword, offset = 0, limit = 10, fields = NULL)
}
\arguments{
\item{keyword}{string with keywords to search for, supports AND and OR and
inclusion and exclusion of terms, for example "+Epidemic +Modeling +Canada -COVID"}

\item{offset}{integer paging offset}

\item{limit}{integer paging length}

\item{fields}{extra fields to include, for example "title,authors"}
}
\value{
list representing paper objects

list representing paper objects
}
\description{
This function retrieves Semantic Scholar data for
a keyword search

This function retrieves Semantic Scholar data for
a keyword search
}
\examples{
\dontrun{
 search_papers(keyword = "literature graph")
 }
\dontrun{
 search_papers(keyword = "literature graph")
 }
}
