% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-semanticscholar.R
\name{semanticscholar-package}
\alias{semanticscholar-package}
\alias{semanticscholar}
\title{Data from Semantic Scholar for use in R}
\description{
This package provides on-demand data access to \href{https://semanticscholar.org}{Semantic Scholar}
through a lightweight API for author and publication data.
}
\keyword{package}
