% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_references.R
\name{zotero_references}
\alias{zotero_references}
\title{References for zotero}
\usage{
zotero_references(identifiers)
}
\arguments{
\item{identifiers}{a character vector of identifiers}
}
\value{
a list with results
}
\description{
Given a vector of identifiers, extract data for use in Zotero
}
\details{
See https://github.com/giocomai/zoteroR for working with Zotero from R
}
\examples{
\dontrun{
if(interactive()){
 ids <- c("10.1038/nrn3241", "CorpusID:37220927")
 zotero_references(ids)
 }
}
}
\seealso{
\code{\link[humaniformat]{parse_names}}
}
